+++ package.json
@@ -46,9 +46,11 @@
   },
   "dependencies": {
     "axios": "^0.16.1",
+    "classnames": "^2.2.5",
     "history": "^4.6.1",
     "lodash": "^4.17.4",
     "moment": "^2.18.1",
+    "prop-types": "^15.5.10",
     "react": "^15.5.4",
     "react-bootstrap": "^0.31.0",
     "react-dom": "^15.5.4",
