+++ public/index.html
@@ -18,6 +18,10 @@
       integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN"
       crossorigin="anonymous"
     >
+    <link
+      href="//fonts.googleapis.com/css?family=Lato"
+      rel="stylesheet"
+    >
 
     <title>Flashiro</title>
   </head>
