+++ src/App.css
@@ -1,0 +1,14 @@
+body,
+html,
+#root {
+  height: 100%;
+}
+
+body {
+  font-family: 'Lato', sans-serif;
+}
+
+.app__wrapper {
+  background-image: url('/background.png');
+  height: 100%;
+}
