+++ src/App.js
@@ -1,12 +1,16 @@
 import React, {Component} from 'react';
 import {Grid} from 'react-bootstrap';
 
+import './App.css';
+
 class App extends Component {
   render() {
     return (
-      <Grid>
-        {this.props.children}
-      </Grid>
+      <div className="app__wrapper">
+        <Grid>
+          {this.props.children}
+        </Grid>
+      </div>
     );
   }
 }
