--- src/components/dashboard.css
@@ -1,6 +1,0 @@
-.dashboard-tile {
-  background-color: #ecf0f1;
-  border-radius: 4px;
-  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, .75);
-  padding: 10px 20px;
-}
